refactor(web): authenticateのバレルをUIと本体に分割
shared な AppSidebar/AppHeader が Public API バレル経由で LogoutButton を
取り込むと、バレルが再export する useSession（swr）や検証ロジック一式が
Storybook のブラウザバンドルへ読み込まれ、CJS interop・カバレッジ母数の
汚染・bundle 肥大を引き起こしていた。

UI 専用の軽量バレル（ui.ts: AuthenticateForm / LogoutButton）を分離し、
shared 層と login/signup ページは UI を ui.ts から取得する。hook・検証・
エラーメッセージ・turnstile は従来どおり index.ts から取得する。

これにより暫定対応として加えた optimizeDeps の swr / pino も不要となるため除去する。
