+++ application/packages/web/src/client/routes/login/page.tsx
@@ -1,4 +1,5 @@
-import { AuthenticateForm, type AuthenticateErrors } from '@/client/features/authenticate'
+import type { AuthenticateErrors } from '@/client/features/authenticate'
+import { AuthenticateForm } from '@/client/features/authenticate/ui'
 import { Card, CardContent, CardFooter, CardHeader, CardTitle } from '../../components/shadcn/card'
 import Footer from '../../components/ui/layout/footer'
 import LandingHeader from '../../components/ui/layout/landing-header'
