+++ application/packages/web/vitest.config.ts
@@ -57,9 +57,7 @@ export default defineConfig(() => {
               'vaul',
               'neverthrow',
               'zod',
-              'swr',
               'swr/mutation',
-              'pino',
               'hono/client',
               'storybook/test',
               'storybook/preview-api',
